refactor: rating stars component
