+++ packages/shared/src/components/cards/FeedSurveyCard.tsx
@@ -1,15 +1,12 @@
 import React, { ReactElement, ReactNode, useEffect, useState } from 'react';
-import classNames from 'classnames';
 import { Card as CardV1 } from './v1/Card';
 import { Card } from './Card';
 import { useFeedLayout } from '../../hooks';
-import { StarIcon } from '../icons';
-import { IconSize } from '../Icon';
 import { Button, ButtonSize, ButtonVariant } from '../buttons/Button';
-import ConditionalWrapper from '../ConditionalWrapper';
 import { useAnalyticsContext } from '../../contexts/AnalyticsContext';
 import { AnalyticsEvent, TargetId, TargetType } from '../../lib/analytics';
 import { updateFeedFeedbackReminder } from '../../graphql/alerts';
+import { RatingStars } from '../utilities/RatingStars';
 
 interface FeedSurveyCardProps {
   title: string;
@@ -29,11 +26,8 @@ export function FeedSurveyCard({
   postFeedbackMessage = 'Thank you for your feedback!',
 }: FeedSurveyCardProps): ReactElement {
   const [score, setScore] = useState(0);
-  const [hovered, setHovered] = useState(0);
   const [submitted, setSubmitted] = useState(false);
-  const [justClicked, setJustClicked] = useState(false);
   const { shouldUseListFeedLayout } = useFeedLayout();
-  const highlighted = hovered || score;
   const CardComponent = shouldUseListFeedLayout ? CardV1 : Card;
   const { trackEvent } = useAnalyticsContext();
 
@@ -55,14 +49,6 @@ export function FeedSurveyCard({
     trackSurveyEvent(AnalyticsEvent.Click, { value: score });
   };
 
-  const onStarClick = (value) => {
-    if (submitted) {
-      return;
-    }
-
-    setScore(value);
-  };
-
   const onHide = () => {
     updateFeedFeedbackReminder();
     trackSurveyEvent(AnalyticsEvent.DismissPromotion);
@@ -74,11 +60,6 @@ export function FeedSurveyCard({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const onOut = () => {
-    setHovered(0);
-    setJustClicked(false);
-  };
-
   return (
     <CardComponent
       className="relative items-center gap-4 overflow-hidden pt-20"
@@ -88,42 +69,7 @@ export function FeedSurveyCard({
       }}
     >
       <h3 className="font-bold typo-title3">{title}</h3>
-      <span className="flex flex-row">
-        {Object.keys([...Array(max)]).map((key, i) => (
-          <ConditionalWrapper
-            key={key}
-            condition={!submitted}
-            wrapper={(component) => (
-              <button
-                key={key}
-                type="button"
-                onClick={() => {
-                  onStarClick(i + 1);
-                  setJustClicked(true);
-                }}
-                onMouseOver={() => setHovered(i + 1)}
-                onFocus={() => setHovered(i + 1)}
-                onMouseOut={onOut}
-                onBlur={onOut}
-              >
-                {component}
-              </button>
-            )}
-          >
-            <StarIcon
-              size={IconSize.XLarge}
-              secondary={i < highlighted}
-              className={classNames(
-                'mx-0.5',
-                i < highlighted
-                  ? 'text-accent-cheese-default'
-                  : 'text-text-secondary',
-                i < highlighted && hovered && !justClicked && 'opacity-[0.8]',
-              )}
-            />
-          </ConditionalWrapper>
-        ))}
-      </span>
+      <RatingStars max={max} isDisabled={submitted} onStarClick={setScore} />
       {submitted && (
         <p className="mt-2 text-center text-text-secondary typo-body">
           <span className="font-bold">{postFeedbackMessage}</span>
